requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
schedule==1.2.2
Werkzeug==3.1.3
